pyproject.toml
[project]
name = "pyra"
version = "0.1.0"
description = "Modern Python IRC bot inspired by Sopel and Eggdrop"
authors = [{name = "Jarsky", email = "jarskynz@gmail.com"}]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "aiohttp>=3.9",
    "aiofiles>=23.0",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "bcrypt>=4.0",
    "loguru>=0.7",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "simpleeval>=0.9",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7",
    "aiosqlite>=0.20",
    "websockets>=12.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29"]
mysql = ["aiomysql>=0.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27",
]

[project.scripts]
pybot = "pybot.__main__:main"
pybot-setup = "pybot.scripts.setup:main"
pybot-ctl = "pybot.scripts.ctl:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["pybot*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "S"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "B"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
